+++ common/base_cpp/exception.h
@@ -70,7 +70,7 @@ public                                                                          
     {                                                                                        \
         va_list args;                                                                        \
         va_start(args, format);                                                              \
-        const size_t len = static_cast<size_t>(strlen(_message));                            \
+        const size_t len = strlen(_message);                                                 \
         vsnprintf(_message + len, sizeof(_message) - len, format, args);                     \
         va_end(args);                                                                        \
     }                                                                                        \
